Create README.md
Welcome to the official repository of the Mental Health and Psychiatric Nursing Science Institute (MHPNSI)-,a pioneering institution dedicated to pioneering advancement in Mental health and wellness programme research, community collaboration and education. Our mission is to empower mental health nurses and professionals with cutting -edge know-how, skills as well as compassionate approaches that positively and pleasantly transform lives.